Move stake tests to a separate directory
