+++ solidity/test/token_stake/TestTokenStake.js
@@ -1,7 +1,7 @@
-import mineBlocks from './helpers/mineBlocks';
-import latestBlock from './helpers/latestBlock';
-import expectThrowWithMessage from './helpers/expectThrowWithMessage'
-import {createSnapshot, restoreSnapshot} from "./helpers/snapshot"
+import mineBlocks from '../helpers/mineBlocks';
+import latestBlock from '../helpers/latestBlock';
+import expectThrowWithMessage from '../helpers/expectThrowWithMessage'
+import {createSnapshot, restoreSnapshot} from "../helpers/snapshot"
 
 const BN = web3.utils.BN
 const chai = require('chai')
